add confirm delete

+++ src/main/java/com/ndb/auction/resolver/UserResolver.java
@@ -3,7 +3,7 @@ package com.ndb.auction.resolver;
 import java.util.List;
 
 import com.ndb.auction.models.GeoLocation;
-import com.ndb.auction.models.user.User;
+import com.ndb.auction.models.User;
 import com.ndb.auction.service.UserDetailsImpl;
 
 import org.springframework.security.access.prepost.PreAuthorize;
@@ -65,11 +65,8 @@ public class UserResolver extends BaseResolver implements GraphQLQueryResolver, 
 
         String rPassword = userService.getRandomPassword(10);
         String encoded = userService.encodePassword(rPassword);
-        user=new User();
-        user.setEmail(email);
-        user.setPassword(encoded);
-        user.setCountry(country);
-        
+        user = new User(email, encoded, country, true);
+
         user.setAvatarPrefix(avatarName);
         user.setAvatarName(shortName);
 
@@ -99,10 +96,19 @@ public class UserResolver extends BaseResolver implements GraphQLQueryResolver, 
     }
 
     @PreAuthorize("isAuthenticated()")
-    public String deleteAccount() {
-        UserDetailsImpl userDetails = (UserDetailsImpl)SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-        String id = userDetails.getId();
-        return userService.deleteUser(id);
+    public String deleteAccount() {        
+        return "To delete your account, please withdraw all your assets from NDB Wallet. Please note deleting process is irreversible.";
+    }
+
+    @PreAuthorize("isAuthenticated()")
+    public String confirmDeleteAccount(String text) {
+        if (text.equals("delete")) {
+            UserDetailsImpl userDetails = (UserDetailsImpl)SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+            String id = userDetails.getId();
+            return userService.deleteUser(id);
+        } else {
+            return "failed";
+        }
     }
 
 //    @PreAuthorize("isAuthenticated")
